Move Redis to own file
